requirements.txt
pandas
numpy
nltk
scikit-learn
google-generativeai
requests
flask
